Character配置模块测试通过
