+++ jyx2/Assets/Scripts/Jyx2UIScripts/ChatUIPanel.cs
@@ -178,7 +178,6 @@ public partial class ChatUIPanel : Jyx2_UIBase
 		}
 
 		//再从人物库找
-		//var role = GameConfigDatabase.Instance.Get<Jyx2ConfigCharacter>(headId);
 		var roleName = LuaToCsBridge.CharacterTable[headId].Name;
 		return roleName;
 	}
